feat: datasets in the tensorlake SDK (#44)
This PR extends the Document AI client with the concept of datasets.

A dataset is a collection of jobs that share the same settings. One can
create and retrieve a dataset and extend it with more files.

An example is added showing how to add files to a dataset and retrieve
its outputs.