Make a more strict GPU health check using pytorch
It turns out that nvidia-smi doesn't detect the known issue
https://github.com/NVIDIA/nvidia-container-toolkit/issues/857

So in addition to that I'm adding a simple pytorch CUDA computation
as a health check. This is known to detect the known issue.

Also refactored the code to extra the health check handler into
a separate class because it quite complex now.