Don't use `which -s` flag for GPU health check (#65)
`-s` flag suppresses `which` output but it's not available in all Linux
distributions, see example error:

```
Usage: /usr/bin/which [-a] args
Illegal option -s
```

Lack of -s flag results in GPU health check not being enabled when GPU
is available. So this is a bug.

Instead of using `-s` flag run which with /dev/null (suppressed) stdout,
stderr.