Remove dep for asyncio stdlib (#72)
Asyncio is part of the stdlib and doesn't need to be used via a dep
package.

This package was setup for python 3.3.

I ran into an issue where asyncio code with fastapi was interacting with
this dep code path from the asyncio package. The fastapi server we're
running is a part of `arxivsearch-client`.