Create a new client for async calls (#99)
If we want to use a sync call with the async method we need to make sure
we create a new client for each sync call.

This will allow a user to do what is natural with sync methods with a
for-loop,
```
for _ in range(10):
  dataset.ingest(...)
```

Without this fix the for loop above will fail. This is because we are
attempting to re-use the same event loop between calls because we share
the client between them. We close the event loop resource when the first
call is done, failing the subsequent call.

This PR is one way to fix it.

Another way to fix it would be to accept a list of inputs for the sync
method and let the sdk handle the concurrency.