Run latest OSS Indexify in Tensorlake SDK CI
We are currently using stale not supported default Executor image
in CI and CI is always red now because of this.

Addressing this issue by running latest published Indexify Server
and dev Executor when testing the latest SDK. The tests tensorlake SDK
package in reinstalled on top of the one coming with the latest
published Indexify Executor. This allows us to test for backward
compatibility of Tensorlake SDK changes that we're adding.

The forward compatibility is tested when changes supporting them are
added to OSS Executor in Indexify repo.

The only test that is currently failing is one of graph update tests.
I tried to fix it but it looks like there's a new regression now
so I commented it out until it gets fixed on Server side.