Use stable Indexify in Tensorlake SDK CI (#161)
These versions are used in Prod Doc AI cluster.
Use them in CI for now to ensure that all the Tensorlake SDK changes are
compatible with the cluster.

We're going to publish Indexify images and packages with dynamic
scheduling enabled so if we stay on latest versions then we'll start
using them in Tensorlake SDK CI which might result in not caught
compatibility issues with the current prod cluster in Tensorlake SDK CI.