Add cache keys for functions and start node parameters (#110)
This change adds keys for cache access checks to graph functions and the
start node's parameters.
For graph functions, these are attached to the graph's JSON definition;
for the start node's parameters, the key is computed by hashing the
serialized input.

This change does _not_ take steps to ensure a deterministic
serialization for the start node's parameters; if/when that becomes
useful, we can add it separately.

Fixes tensorlakeai/indexify#1339