Remove/fix legacy image build features (#180)
* Remove Indexify from the function images. This can lead to installing
(e.g. upgrading to) tensorlake package version which is not compatible
with the one used by the user who builds the image.
* Install tensorlake package with exactly the same version as
`tensorlake deploy` command is using to ensure compatibility with what
the user who builds the image is using.
* Don't allows installing indexify/tensorlake from source into the
image. The use case for this is not clear. We're not using it even for
testing in Platform Executor.

Tested with `tensorlake deploy ./reference_graph.py`