Small refactorings for Function Executor + a couple changes (#233)
This is to prepare FE for new features.

The behavioral changes added:

* Don't log customers' initialize exceptions as this is customer data.
We store these exceptions in blob store now. This is why we don't need
to log them.
* Pass health checks only after successful FE initialization. This
allows to simplify FE code a bit and is a fine thing to require. We
already run health checks only after successful initialization.