feat: Datasets V2 (#225)
## Description

Revamps and fixes a lot of issues with submitting parse operations and
adds the capacity to use the V2 iteration of Datasets.

At the moment, the tests are only running on workflow dispatch and merge
on main.

### Test run

```
From test_parse:
.
----------------------------------------------------------------------
Ran 4 tests in 111.047s
OK
```

```
From test_dataset:
.
----------------------------------------------------------------------
Ran 4 tests in 131.502s

OK
```