fix: re-structure models module (#259)
### Description

We have to move `PageFragmentType` or copy the enum to the `options`
module. This is because having a reference to the `results` module
creates a circular module dependency between `options` and `results`.

Following @diptanu's advice, this PR privatizes the inner modules of the
`models` module, and adds them to the root`__init__.py` to help the
transition.

**DO NOT MERGE THIS** until every Jupyter notebook is migrated to use
the root module as import instead of the submodules.