Derive image builder service url from indexify url (#363)
I deleted this logic recently but without it when ppl want to use dev
cluster they have now to set two env vars:
- INDEXIFY_URL
- TENSORLAKE_BUILD_SERVICE

Previously we only needed to set INDEXIFY_URL in this case.
This is inconvenient. This also breaks our existing testing infra in
dev. So we need to bring this logic back until we have a cleaner
approach.