Move RequestContext out of function parameters to RequestContext.get()
The approach with RequestContext in function parameters required
passing the ctx object through every function that uses and not
uses it if a downstream function uses it.

Instead of passing RequestContext in function parameters we're
providing a new getter called RequestContext.get() to get
context of the current request.

RequestContext.get() fails if called from a thread spawned by
a tensorlake function call. In all other cases including
async functions spawning async tasks it works. A clear exception
is raised if RequestContext.get() is called outside of the main
Tensorlake Function thread.