Use CloudEvent as the standard format for internal logs and events.
We want to have a consistent format for everything that happens inside
of an executor.

This change adds a new cloud_events module that turns internal
events into a CloudEvent object. It also adds utilities to
print logs with the CloudEvent format, and to push events to
an external collector.