Resolve the cloud URL based on the API url. (#433)
If the API url is set to https://api.tensorlake.dev, we want the cloud
url to be automatically set to https://cloud.tensorlake.dev without
having to specify that in the configuration.

This change checks the API url to resolve the right cloud url. That way,
when you run `tensorlake login` with an instance that's not production,
you can get redirected to the right cloud interface directly, otherwise
you always end up in https://cloud.tensorlake.ai.

---------

Co-authored-by: Raees Iqbal <10067728+RaeesBhatti@users.noreply.github.com>