Remove retry python package from dependencies (#471)
retry package is not maintained (last release is in 2016), see
https://pypi.org/project/retry/. The package depends on py package which
is also in deep maintanence mode, see https://pypi.org/project/py. The
latest version of py has a vulnerability. We also need to reduce the
number of dependencies in tensorlake package to not force ppl to use
fixed package versions in their function images.

This is why it's better to remove retry package dependency and use our
internal implementation of retries that we already have.

I also extended our owb retry package with a sensible default to retry
all Exceptions. This is because by default without this our
implementation won't retry anything and thus in every place where we use
our implementation of retries we had to specify the exception classes
manually. I think this default is better cause it's easy to miss it and
end up not retrying anything at all.