IPC request context for remote mode (Function Executor)
Each FE creates an HTTP server and an httpx.Client for it
on FE initialization. Then both are reused during FE
lifetime to call and implement RequestContext methods
for all allocations running in the FE.

Also unify request progress and metrics printing code in FE
and local mode. This helps keeping the printed messages
consistent in both modes.

Also stop sending user metrics to executor. We'll use the
same approach as we use for user logs: print them as cloud events.
Later if we want we can use otel for this.