Use content derived IDs for Awaitables in FE
This is required for durability/replayability feature.

Awaitable IDs supporting durable execution are the same across different executions (allocations) of the same parent function call
if the parent function call is deterministic, i.e. it creates the same awaitables in the same order each time it's executed.
If this is not the case then the Awaitable IDs will differ between executions, which may lead to re-execution of some function calls
even if their inputs are the same as in a previous execution. If re-execution is allowed or not will be controlled by Server.

To produce Awaitable IDs supporting durable execution, we compute them as a hash of:
- parent_function_call_id
- child_call_sequence_number
- Awaitable-specific metadata that is not expected to change between executions of the same parent function call.

We're deliberately not hashing entire awaitables to produce their durable IDs. This is because hashing entire awaitables
is an expensive operation (i.e. hashing gigabytes of function call parameters) but this is also inconvenient for users
in many cases, i.e.:
- Seamless Schema Evolution: Users may want to change the schema of function parameters (e.g. add a new field with a default value to a pydantic model).
- Non-deterministic functions: Users may want to use non-deterministic functions (e.g. functions that read current time or random values)
  inside otherwise deterministic function call trees.
- Avoiding "Serialization Flakiness" - Strict equality checks on serialized data are fragile and can lead to false positive
  re-executions due to minor, non-semantic changes in serialization (e.g. different field ordering in protobufs, or insertion order in dicts).
- Decoupling Logic from Data. We adhere to a philosophy of being Strict on Control Flow but Lenient on Data. "The History is the Source of Truth."